don't build readv for windows
