+++ common/buf/readv_reader_windows.go
@@ -1,0 +1,13 @@
+// +build windows
+package buf
+
+import (
+	"io"
+	"syscall"
+)
+
+var useReadv = false
+
+func NewReadVReader(reader io.Reader, rawConn syscall.RawConn) *ReadVReader {
+	return nil
+}
